requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
feedparser==6.0.11
pyyaml==6.0.2
jinja2==3.1.5
python-multipart==0.0.20
